rust(parsers): propagate BigInt through all 9 frontend parsers
Wire all BigIntLiteral construction sites in the SWC TypeScript
parser, Solidity, Move, Java, Python, Ruby, Zig, Go DSL, and Rust
DSL parsers to emit num_bigint::BigInt values. The SWC parser now
clones the BigInt directly out of the swc_ecma_ast Lit::BigInt token
rather than funneling it through i128::from_str (which truncated
256-bit values to 0).

The Java parser also gets a new parse_radix_to_bigint helper to
capture full-precision decimal/hex literals from BigInteger.valueOf
/ Bigint.of calls. expand_fixed_arrays.rs's as_literal_index now
returns Option<i128> via BigInt::to_i128() so oversize indexes are
correctly rejected at the (small-int) index-bound check rather than
silently truncated.

+++ compilers/rust/src/frontend/parser_gocontract.rs
@@ -36,6 +36,7 @@
 //! - `:=` -> VariableDecl (mutable)
 //! - standalone functions (no receiver) -> private helper methods
 
+use num_bigint::BigInt;
 use super::ast::{
     BinaryOp, ContractNode, Expression, MethodNode, ParamNode, PrimitiveTypeName, PropertyNode,
     SourceLocation, Statement, TypeNode, UnaryOp, Visibility,
@@ -1430,7 +1431,7 @@ impl<'a> GoParser<'a> {
                 name: "_i".to_string(),
                 var_type: None,
                 mutable: true,
-                init: Expression::BigIntLiteral { value: 0 },
+                init: Expression::BigIntLiteral { value: BigInt::from(0) },
                 source_location: loc.clone(),
             };
             let dummy_update = Statement::ExpressionStatement {
@@ -1736,7 +1737,7 @@ impl<'a> GoParser<'a> {
         match self.current().typ.clone() {
             TokenType::Number(n) => {
                 self.advance();
-                Some(Expression::BigIntLiteral { value: n })
+                Some(Expression::BigIntLiteral { value: BigInt::from(n) })
             }
 
             TokenType::True => {
