+++ compilers/rust/src/frontend/parser_move.rs
@@ -27,6 +27,7 @@
 //! - snake_case identifiers -> camelCase in AST
 //! - Move builtins mapped: `check_sig` -> `checkSig`, `hash160` -> `hash160`, etc.
 
+use num_bigint::BigInt;
 use super::ast::{
     BinaryOp, ContractNode, Expression, MethodNode, ParamNode, PrimitiveTypeName, PropertyNode,
     SourceLocation, Statement, TypeNode, UnaryOp, Visibility,
@@ -1286,12 +1287,12 @@ impl<'a> MoveParser<'a> {
                 name: "_w".to_string(),
                 var_type: None,
                 mutable: true,
-                init: Expression::BigIntLiteral { value: 0 },
+                init: Expression::BigIntLiteral { value: BigInt::from(0) },
                 source_location: self.loc(),
             }),
             condition,
             update: Box::new(Statement::ExpressionStatement {
-                expression: Expression::BigIntLiteral { value: 0 },
+                expression: Expression::BigIntLiteral { value: BigInt::from(0) },
                 source_location: self.loc(),
             }),
             body,
@@ -1712,7 +1713,7 @@ impl<'a> MoveParser<'a> {
 
     fn parse_primary(&mut self) -> Expression {
         match self.advance() {
-            Token::NumberLit(v) => Expression::BigIntLiteral { value: v },
+            Token::NumberLit(v) => Expression::BigIntLiteral { value: BigInt::from(v) },
             Token::True => Expression::BoolLiteral { value: true },
             Token::False => Expression::BoolLiteral { value: false },
             Token::StringLit(v) => Expression::ByteStringLiteral { value: v },
@@ -1738,7 +1739,7 @@ impl<'a> MoveParser<'a> {
             other => {
                 self.errors
                     .push(Diagnostic::error(format!("Unexpected token in expression: {:?}", other), None));
-                Expression::BigIntLiteral { value: 0 }
+                Expression::BigIntLiteral { value: BigInt::from(0) }
             }
         }
     }
